+++ filament/src/PostProcessManager.h
@@ -37,6 +37,7 @@
 #include <filament/Viewport.h>
 
 #include <private/filament/EngineEnums.h>
+#include <private/filament/Variant.h>
 
 #include <backend/DriverEnums.h>
 #include <backend/Handle.h>
@@ -215,9 +216,15 @@ public:
             FrameGraphId<FrameGraphTexture> inout) noexcept;
 
     // clear depth buffer pass
-    void clearAncillaryBuffersPrepare(backend::DriverApi& driver) noexcept;
+    void clearAncillaryBuffersPrepare(backend::DriverApi& driver,
+            Variant::type_t variant) noexcept;
     void clearAncillaryBuffers(backend::DriverApi& driver,
-            backend::TargetBufferFlags attachments) const noexcept;
+            backend::TargetBufferFlags attachments,
+            Variant::type_t variant) const noexcept;
+
+    // postfx fog
+    void fogPrepare(backend::DriverApi& driver) noexcept;
+    void fog(backend::DriverApi& driver) noexcept;
 
     // Anti-aliasing
     FrameGraphId<FrameGraphTexture> fxaa(FrameGraph& fg,
@@ -351,7 +358,12 @@ public:
         void terminate(FEngine& engine) noexcept;
 
         FMaterial* getMaterial(FEngine& engine, backend::DriverApi& driver,
-                PostProcessVariant variant = PostProcessVariant::OPAQUE) const noexcept;
+                Variant::type_t variant) const noexcept;
+
+        FMaterial* getMaterial(FEngine& engine, backend::DriverApi& driver,
+                PostProcessVariant variant = PostProcessVariant::OPAQUE) const noexcept {
+                return getMaterial(engine, driver, Variant::type_t(variant));
+        }
 
     private:
         void loadMaterial(FEngine& engine) const noexcept;
@@ -376,9 +388,12 @@ public:
 
     void bindPostProcessDescriptorSet(backend::DriverApi& driver) const noexcept;
 
-    backend::PipelineState getPipelineState(
-            FMaterial const* ma,
-            PostProcessVariant variant = PostProcessVariant::OPAQUE) const noexcept;
+    backend::PipelineState getPipelineState(FMaterial const* ma, Variant::type_t variant) const noexcept;
+
+    backend::PipelineState getPipelineState(FMaterial const* ma,
+                    PostProcessVariant variant = PostProcessVariant::OPAQUE) const noexcept {
+            return getPipelineState(ma, Variant::type_t(variant));
+    }
 
     void renderFullScreenQuad(FrameGraphResources::RenderPassInfo const& out,
             backend::PipelineState const& pipeline,