+++ web/filament-js/jsbindings.cpp
@@ -250,6 +250,7 @@ DecodedImage decodeImage(BufferDescriptor encoded_data, int requested_ncomp) {
 
 EMSCRIPTEN_BINDINGS(jsbindings) {
 
+
 // MATH TYPES
 // ----------
 // Individual JavaScript objects for math types would be too heavy, so instead we simply accept
@@ -310,6 +311,28 @@ value_object<Box>("Box")
 value_object<filament::Aabb>("Aabb")
     .field("min", &filament::Aabb::min)
     .field("max", &filament::Aabb::max);
+
+value_object<filament::Engine::Config>("Engine$Config")
+    .field("commandBufferSizeMB", &filament::Engine::Config::commandBufferSizeMB)
+    .field("perRenderPassArenaSizeMB", &filament::Engine::Config::perRenderPassArenaSizeMB)
+    .field("driverHandleArenaSizeMB", &filament::Engine::Config::driverHandleArenaSizeMB)
+    .field("minCommandBufferSizeMB", &filament::Engine::Config::minCommandBufferSizeMB)
+    .field("perFrameCommandsSizeMB", &filament::Engine::Config::perFrameCommandsSizeMB)
+    .field("jobSystemThreadCount", &filament::Engine::Config::jobSystemThreadCount)
+    .field("metalUploadBufferSizeBytes", &filament::Engine::Config::metalUploadBufferSizeBytes)
+    .field("metalDisablePanicOnDrawableFailure", &filament::Engine::Config::metalDisablePanicOnDrawableFailure)
+    .field("disableParallelShaderCompile", &filament::Engine::Config::disableParallelShaderCompile)
+    .field("stereoscopicType", &filament::Engine::Config::stereoscopicType)
+    .field("stereoscopicEyeCount", &filament::Engine::Config::stereoscopicEyeCount)
+    .field("resourceAllocatorCacheSizeMB", &filament::Engine::Config::resourceAllocatorCacheSizeMB)
+    .field("resourceAllocatorCacheMaxAge", &filament::Engine::Config::resourceAllocatorCacheMaxAge)
+    .field("disableHandleUseAfterFreeCheck", &filament::Engine::Config::disableHandleUseAfterFreeCheck)
+    .field("preferredShaderLanguage", &filament::Engine::Config::preferredShaderLanguage)
+    .field("forceGLES2Context", &filament::Engine::Config::forceGLES2Context)
+    .field("assertNativeWindowIsValid", &filament::Engine::Config::assertNativeWindowIsValid)
+    .field("gpuContextPriority", &filament::Engine::Config::gpuContextPriority)
+    .field("sharedUboInitialSizeInBytes", &filament::Engine::Config::sharedUboInitialSizeInBytes)
+    .field("asynchronousMode", &filament::Engine::Config::asynchronousMode);
 
 value_object<filament::Renderer::ClearOptions>("Renderer$ClearOptions")
     .field("clearColor", &filament::Renderer::ClearOptions::clearColor)
@@ -375,9 +398,31 @@ register_vector<allow_raw_pointer<MaterialInstance*>>("MaterialInstanceVector");
 // CORE FILAMENT CLASSES
 // ---------------------
 
+enum_<Engine::StereoscopicType>("StereoscopicType")
+    .value("NONE", Engine::StereoscopicType::NONE)
+    .value("INSTANCED", Engine::StereoscopicType::INSTANCED)
+    .value("MULTIVIEW", Engine::StereoscopicType::MULTIVIEW);
+
+enum_<Engine::GpuContextPriority>("GpuContextPriority")
+    .value("DEFAULT", Engine::GpuContextPriority::DEFAULT)
+    .value("LOW", Engine::GpuContextPriority::LOW)
+    .value("MEDIUM", Engine::GpuContextPriority::MEDIUM)
+    .value("HIGH", Engine::GpuContextPriority::HIGH)
+    .value("REALTIME", Engine::GpuContextPriority::REALTIME);
+
+enum_<Engine::AsynchronousMode>("AsynchronousMode")
+    .value("NONE", Engine::AsynchronousMode::NONE)
+    .value("THREAD_PREFERRED", Engine::AsynchronousMode::THREAD_PREFERRED)
+    .value("AMORTIZATION", Engine::AsynchronousMode::AMORTIZATION);
+
+enum_<Engine::Config::ShaderLanguage>("ShaderLanguage")
+    .value("DEFAULT", Engine::Config::ShaderLanguage::DEFAULT)
+    .value("MSL", Engine::Config::ShaderLanguage::MSL)
+    .value("METAL_LIBRARY", Engine::Config::ShaderLanguage::METAL_LIBRARY);
+
 /// Engine ::core class:: Central manager and resource owner.
 class_<Engine>("Engine")
-    .class_function("_create", (Engine* (*)()) [] {
+    .class_function("_create", (Engine* (*)(Engine::Config)) [] (Engine::Config config) {
         EM_ASM_INT({
             const options = window.filament_glOptions;
             const context = window.filament_glContext;
@@ -385,8 +430,15 @@ class_<Engine>("Engine")
             window.filament_contextHandle = handle;
             GL.makeContextCurrent(handle);
         });
-        return Engine::create();
+        return Engine::create(Engine::Backend::DEFAULT, nullptr, nullptr, &config);
     }, allow_raw_pointers())
+
+    // Create a default Engine configuration. This is for internal use to ensure that engine
+    // creation logic in 'extensions.js' does not have to populate the default configuration
+    // variables manually.
+    .class_function("createDefaultConfig", (Engine::Config (*)()) [] {
+        return Engine::Config();
+    })
 
     .class_function("getSteadyClockTimeNano", &Engine::getSteadyClockTimeNano)
 
@@ -404,7 +456,11 @@ class_<Engine>("Engine")
 
     .function("getActiveFeatureLevel", &Engine::getActiveFeatureLevel)
 
+    .function("getBackend", &Engine::getBackend)
+
     .class_function("getMaxStereoscopicEyes", &Engine::getMaxStereoscopicEyes)
+
+    .function("getConfig", &Engine::getConfig)
 
     .function("_execute", EMBIND_LAMBDA(void, (Engine* engine), {
         EM_ASM_INT({
@@ -678,6 +734,13 @@ class_<View>("View")
     .function("_setTemporalAntiAliasingOptions", &View::setTemporalAntiAliasingOptions)
     .function("_setScreenSpaceReflectionsOptions", &View::setScreenSpaceReflectionsOptions)
     .function("_setBloomOptions", &View::setBloomOptions)
+    .function("setShadowingEnabled", &View::setShadowingEnabled)
+    .function("setFrontFaceWindingInverted", &View::setFrontFaceWindingInverted)
+    .function("isFrontFaceWindingInverted", &View::isFrontFaceWindingInverted)
+    .function("setDynamicLightingOptions", &View::setDynamicLightingOptions)
+    .function("setRenderQuality", &View::setRenderQuality)
+    .function("setDynamicResolutionOptions", &View::setDynamicResolutionOptions)
+    .function("getDynamicResolutionOptions", &View::getDynamicResolutionOptions)
     .function("_setFogOptions", &View::setFogOptions)
     .function("_setVignetteOptions", &View::setVignetteOptions)
     .function("_setGuardBandOptions", &View::setGuardBandOptions)
@@ -777,6 +840,8 @@ class_<Camera>("Camera")
     }), allow_raw_pointers())
 
     .function("getScaling", &Camera::getScaling)
+    .function("setShift", &Camera::setShift)
+    .function("getShift", &Camera::getShift)
 
     .function("getNear", &Camera::getNear)
     .function("getCullingFar", &Camera::getCullingFar)
@@ -937,7 +1002,9 @@ class_<RenderTarget>("RenderTarget")
     })
     .function("getMipLevel", &RenderTarget::getMipLevel)
     .function("getFace", &RenderTarget::getFace)
-    .function("getLayer", &RenderTarget::getLayer);
+    .function("getLayer", &RenderTarget::getLayer)
+    .function("getTexture", &RenderTarget::getTexture, allow_raw_pointers())
+    .function("getSupportedColorAttachmentsCount", &RenderTarget::getSupportedColorAttachmentsCount);
 
 class_<RenderableBuilder>("RenderableManager$Builder")
     .BUILDER_FUNCTION("geometry", RenderableBuilder, (RenderableBuilder* builder,
@@ -1068,15 +1135,22 @@ class_<RenderableManager>("RenderableManager")
     .function("setAxisAlignedBoundingBox", &RenderableManager::setAxisAlignedBoundingBox)
     .function("setLayerMask", &RenderableManager::setLayerMask)
     .function("setPriority", &RenderableManager::setPriority)
+    .function("getPriority", &RenderableManager::getPriority)
     .function("setChannel", &RenderableManager::setChannel)
+    .function("getChannel", &RenderableManager::getChannel)
     .function("setCastShadows", &RenderableManager::setCastShadows)
     .function("setReceiveShadows", &RenderableManager::setReceiveShadows)
+    .function("setScreenSpaceContactShadows", &RenderableManager::setScreenSpaceContactShadows)
     .function("isShadowCaster", &RenderableManager::isShadowCaster)
     .function("isShadowReceiver", &RenderableManager::isShadowReceiver)
     .function("setLightChannel", &RenderableManager::setLightChannel)
     .function("getLightChannel", &RenderableManager::getLightChannel)
     .function("setFogEnabled", &RenderableManager::setFogEnabled)
     .function("getFogEnabled", &RenderableManager::getFogEnabled)
+    .function("setCulling", &RenderableManager::setCulling)
+    .function("isCullingEnabled", &RenderableManager::isCullingEnabled)
+    .function("setScreenSpaceContactShadows", &RenderableManager::setScreenSpaceContactShadows)
+    .function("isScreenSpaceContactShadowsEnabled", &RenderableManager::isScreenSpaceContactShadowsEnabled)
 
     .function("setBones", EMBIND_LAMBDA(void, (RenderableManager* self,
             RenderableManager::Instance instance, emscripten::val transforms, size_t offset), {
@@ -1125,8 +1199,10 @@ class_<RenderableManager>("RenderableManager")
     }), allow_raw_pointers())
 
     .function("setBlendOrderAt", &RenderableManager::setBlendOrderAt)
+    .function("getBlendOrderAt", &RenderableManager::getBlendOrderAt)
 
     .function("setGlobalBlendOrderEnabledAt", &RenderableManager::setGlobalBlendOrderEnabledAt)
+    .function("isGlobalBlendOrderEnabledAt", &RenderableManager::isGlobalBlendOrderEnabledAt)
 
     .function("getEnabledAttributesAt", EMBIND_LAMBDA(uint32_t, (RenderableManager* self,
             RenderableManager::Instance instance, size_t primitiveIndex), {
@@ -1590,6 +1666,9 @@ class_<IblBuilder>("IndirectLight$Builder")
 class_<Skybox>("Skybox")
     .class_function("Builder", (SkyBuilder (*)()) [] { return SkyBuilder(); })
     .function("setColor", &Skybox::setColor)
+    .function("setLayerMask", &Skybox::setLayerMask)
+    .function("getLayerMask", &Skybox::getLayerMask)
+    .function("getIntensity", &Skybox::getIntensity)
     .function("getTexture", EMBIND_LAMBDA(Texture*, (Skybox* skybox), {
         return (Texture*) skybox->getTexture(); // cast away const to appease embind
     }), allow_raw_pointers());