+++ RELEASE_NOTES.md
@@ -7,6 +7,14 @@ A new header is inserted each time a *tag* is created.
 Instead, if you are authoring a PR for the main branch, add your release note to
 [NEW_RELEASE_NOTES.md](./NEW_RELEASE_NOTES.md).
 
+## v1.69.3
+
+
+## v1.69.2
+
+- engine: fix shader compilation failure in TAA material
+- engine: fix stereo & parallel shader compilation
+
 ## v1.69.1